backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.1
PyJWT==2.8.0
python-dotenv==1.0.0
Pillow==10.1.0
reportlab==4.0.7
django-filter==23.5
